Remove object variants from auto-assignable fields
The idea is that you **can't know** that a field from an object-variant
will be used to assign to a field on an object.
Might be that the variant of the object variant does not supply fields
for the assignment, then you basically just lied.

This corrects that issue by just forcing the user to deal with assignments on variants themselves.